optimizer-service/requirements.txt
ortools>=9.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
# Phase 3 — production hardening
pyjwt[crypto]>=2.9.0       # Supabase JWT verification (HS256)
slowapi>=0.1.9             # Per-IP rate limiting
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-exporter-otlp>=1.27.0
# Test harness — only needed for `pytest` runs, not the production image.
pytest>=8.0.0
httpx>=0.27.0  # required by fastapi.testclient for /audit tests
